+++ src/entities/PracticalArea/index.tsx
@@ -1,8 +1,42 @@
 
+import { useCallback, useEffect, useState } from 'react';
+import { ArticlesMockService, IArticle } from '../../api/Articles';
 import { isSvg } from '../../shared';
 import { Card, Paragraph, Title } from '../../widgets';
 import styles from './index.module.css';
 export default function PracticalArea() {
+	const [articles, setArticles] = useState<IArticle[] | null>(null);
+
+	const getArticles = useCallback(async () => {
+		try {
+			const response : IArticle[] = await Promise.all([
+				ArticlesMockService.getArticleById('4'),
+				ArticlesMockService.getArticleById('5'),
+				ArticlesMockService.getArticleById('6')
+			]) as IArticle[];
+			setArticles(response);
+		} catch (err) {
+			console.error('Failed to fetch articles:', err);
+			setArticles([]);
+		}
+	}, []);
+
+	// Fetch articles from API and set to state
+	useEffect(() => {
+		getArticles();
+	});
+
+	if (articles === null){
+		return <p>Loading...</p>;
+	};
+
+	if (articles !== null && articles.length == 0){
+		return (
+			<div className={styles['not-found']}>
+				<p>No articles found.</p>
+			</div>
+		);
+	}
 	return (
 		<>
 			<Title text={'Areas of Practice'} className={styles['title']} size='h1'
@@ -11,17 +45,11 @@ export default function PracticalArea() {
             legal issues will produce the best-possible outcome.`}
 			size={24}
 			/>
-			{/**TODO: сделать api */}
+			
 			<div className={styles['main-container']}>
-				<Card id={'1'} iconUrl={{dataSource: '/world-icon.svg', type: isSvg('/world-icon.svg') ? 'svg' : 'image'}} title='Intellectual Property' info={`
-				Intellectual Property law deals with laws to protect creators and owners of inventions, 
-				writing, music, designs and other works.`}/>
-
-				<Card id={'2'} iconUrl={{dataSource: '/estate-icon.svg', type: isSvg('/estate-icon.svg') ? 'svg' : 'image'}} title='Real Estate' info={`
-				Real estate law is a branch of civil law that covers the right to possess, use, and enjoy land.`}/>
-
-				<Card id={'3'} iconUrl={{dataSource: '/tax-law-icon.svg', type: isSvg('/tax-law-icon.svg') ? 'svg' : 'image'}} title='Tax Law' info={`
-				Tax law cover income, corporate, excise, luxury, estate and property taxes, to name a few.`}/>
+				{articles.map((article: IArticle) => {
+					return <Card key={article.id} id={article.id} iconUrl={{dataSource: article.iconUrl.dataSource, type: isSvg(article.iconUrl.dataSource) ? 'svg' : 'image'}} title={article.title} info={article.info}/>;
+				})}
 			</div>
 		</>
 	);